Move ClientRegistrationTest to the new testsuite
Closes #48199
