+++ testsuite/integration-arquillian/tests/base/testsuites/database-suite
@@ -1,7 +1,6 @@
 AccountRestServiceTest
 AuthorizationCodeTest
 AuthorizationTest
-ClientRegistrationTest
 ExportImportTest
 KcOidcBrokerTest
 LDAPUserLoginTest
